Projet CRUD

+++ src/services/api/admin/userapi.js
@@ -1,6 +1,7 @@
 import axios from "axios";
 import { getInStore, TOKEN_KEY } from '../../../services/store'
 import jwt_decode from "jwt-decode";
+import { getAllClientByEmailUser } from "../client/clientapi";
 const BASE_URL = "http://localhost:3333/"
 const header = {headers: { authorization: `Bearer ${getInStore(TOKEN_KEY)}`}}
 
@@ -16,7 +17,8 @@ export async function getAllUser(){
 }
 
 export async function getUserByEmail(email) {
-    const result = await axios.post(BASE_URL + 'user', {email: email})
+    const url = BASE_URL + 'user/' + email
+    const result = await axios.get(url, header)
     .then((response) => response)
     .catch((error) => console.log(error))
 
@@ -40,20 +42,25 @@ export async function insertUser(email, password, nom, prenom, date_de_naissance
     return result
 }
 
-export async function updateUser(id, email, password){
-    const url = BASE_URL + "user/edit/" + id
-    const result = await axios.put(url, {email: email, password: password})
+export async function updateUser(email, password, nom, prenom, date_de_naissance, numero_telephone, adresse_1, adresse_2, code_postal, ville, ca_annuel_max, taux_charge, administrator, emailmodif){
+    const url = BASE_URL + "user/edit/" + emailmodif
+    const result = await axios.put(url, {email: email, password: password, nom: nom, prenom: prenom, date_de_naissance: date_de_naissance, numero_telephone: numero_telephone, adresse_1: adresse_1, adresse_2: adresse_2, code_postal: code_postal, ville: ville, ca_annuel_max: ca_annuel_max, taux_charge: taux_charge, administrator: administrator}, header)
     .then((response) => response)
     .catch((error) => console.log(error))
 
     return result
 }
 
-export async function deleteUser(id){
-    const url = BASE_URL + "user/delete/" + id
-    const result = await axios.delete(url)
-    .then((response) => response)
-    .catch((error) => console.log(error))
+export async function deleteUser(email){
+    const client = await getAllClientByEmailUser(email)
+    if(client.length === 0) {
+        const url = BASE_URL + "user/delete/" + email
+        const result = await axios.delete(url)
+        .then((response) => response)
+        .catch((error) => console.log(error))
 
-    return result
+        return result
+    }else{
+        return false
+    }
 }
